Rename : `/container` 아래 auth, user 폴더 추가
