+++ frontend/src/modules/user/user.js
@@ -1,6 +1,6 @@
 import { createAction, handleActions } from 'redux-actions';
-import createRequestSaga, { createRequestActionTypes } from '../services/createRequestSaga';
-import * as authAPI from '../services/api/authAPI';
+import createRequestSaga, { createRequestActionTypes } from '../../services/createRequestSaga';
+import * as authAPI from '../../services/api/authAPI';
 import { call, takeLatest } from 'redux-saga/effects';
 
 const TEMP_SET_USER = 'user/TEMP_SET_USER'; // 새로고침 이후 임시 로그인 처리
